+++ routes/auth.js
@@ -72,14 +72,29 @@ router.post('/signup',
   } else {    
     return userService.signup(req.body.email, req.body.password, req.body.firstname, req.body.lastname, false)
       .then(function (savedUser) {
-
-
+        
         winston.debug('-- >> -- >> savedUser ', savedUser.toObject());
 
-        if (!req.body.disableEmail){
-          emailService.sendVerifyEmailAddress(savedUser.email, savedUser);
-        }
+        // let skipVerificationEmail = false;
+        // if (req.headers.authorization) {
+
+        //   let token = req.headers.authorization.split(" ")[1];
+        //   let decode = jwt.verify(token, configSecret)
+        //   if (decode && (decode.email === process.env.ADMIN_EMAIL)) {
+        //     skipVerificationEmail = true;
+        //     winston.verbose("skip sending verification email")
+        //   }
+        // }
+
+        // if (!req.body.disableEmail){
+        //   if (!skipVerificationEmail) {
+        //     emailService.sendVerifyEmailAddress(savedUser.email, savedUser);
+        //   }
+        // }
         
+        if (!req.body.disableEmail){
+            emailService.sendVerifyEmailAddress(savedUser.email, savedUser);
+        }
 
 
         /*