+++ routes/project.js
@@ -11,6 +11,7 @@ var operatingHoursService = require("../services/operatingHoursService");
 var winston = require('../config/winston');
 var roleChecker = require('../middleware/has-role');
 
+
 // THE THREE FOLLOWS IMPORTS  ARE USED FOR AUTHENTICATION IN THE ROUTE
 var passport = require('passport');
 require('../middleware/passport')(passport);
@@ -19,6 +20,21 @@ var RoleConstants = require("../models/roleConstants");
 var cacheUtil = require('../utils/cacheUtil');
 var orgUtil = require("../utils/orgUtil");
 var cacheEnabler = require("../services/cacheEnabler");
+
+/**
+ * NEW
+ */
+var jwt = require('jsonwebtoken');
+var config = require('../config/database');
+
+let configSecret = process.env.GLOBAL_SECRET || config.secret;
+var pKey = process.env.GLOBAL_SECRET_OR_PUB_KEY;
+if (pKey) {
+  configSecret = pKey.replace(/\\n/g, '\n');
+}
+/**
+ * End NEW
+ */
 
 router.post('/', [passport.authenticate(['basic', 'jwt'], { session: false }), validtoken], async (req, res) => {
   
@@ -71,10 +87,219 @@ router.delete('/:projectid', [passport.authenticate(['basic', 'jwt'], { session:
   });
 });
 
+// router.put('/:projectid/update', function (req, res) {
+// // router.put('/:projectid/profile', [passport.authenticate(['basic', 'jwt'], { session: false }), validtoken], function (req, res) {
+  
+//   // Get token from header authorization
+//   let token = req.headers.authorization;
+//   token = token.split(" ")[1];
+
+//   try {
+//     let decoded = jwt.verify(token, configSecret)
+//     winston.debug("user decode: ", decoded);
+
+//     if (!process.env.ADMIN_EMAIL) {
+//       winston.warn("Missing admin email parameter in environment");
+//       return res.status(401).send({ success: false, error: "Missing admin email parameter"});
+//     }
+
+//     if (decoded.email !== process.env.ADMIN_EMAIL) {
+//       winston.warn("Profile modification: permission denied.");
+//       return res.status(403).send({ success: false, error: "You don't have the permission required to modify the project profile"});
+//     } 
+
+//     /**
+//      * modify the project profile here
+//      */
+//     var update = {};
+
+//     if (req.body.name!=undefined) {
+//       update.name = req.body.name;
+//     }
+  
+//     if (req.body.activeOperatingHours!=undefined) {
+//       update.activeOperatingHours = req.body.activeOperatingHours;
+//     }
+    
+//     if (req.body.operatingHours!=undefined) {
+//       update.operatingHours = req.body.operatingHours;
+//     }
+    
+//     if (req.body.settings!=undefined) {
+//       update.settings = req.body.settings;
+//     }
+
+//     if (req.body["settings.email.autoSendTranscriptToRequester"]!=undefined) {
+//       update["settings.email.autoSendTranscriptToRequester"] = req.body["settings.email.autoSendTranscriptToRequester"];
+//     }
+//     if (req.body["settings.email.notification.conversation.assigned"]!=undefined) {
+//       update["settings.email.notification.conversation.assigned"] = req.body["settings.email.notification.conversation.assigned"];
+//     }
+//     if (req.body["settings.email.notification.conversation.pooled"]!=undefined) {
+//       update["settings.email.notification.conversation.pooled"] = req.body["settings.email.notification.conversation.pooled"];
+//     }
+//     if (req.body["settings.email.templates.assignedRequest"]!=undefined) {
+//       update["settings.email.templates.assignedRequest"] = req.body["settings.email.templates.assignedRequest"];
+//     }
+//     if (req.body["settings.email.templates.assignedEmailMessage"]!=undefined) {
+//       update["settings.email.templates.assignedEmailMessage"] = req.body["settings.email.templates.assignedEmailMessage"];
+//     }
+//     if (req.body["settings.email.templates.pooledRequest"]!=undefined) {
+//       update["settings.email.templates.pooledRequest"] = req.body["settings.email.templates.pooledRequest"];
+//     }
+//     if (req.body["settings.email.templates.pooledEmailMessage"]!=undefined) {
+//       update["settings.email.templates.pooledEmailMessage"] = req.body["settings.email.templates.pooledEmailMessage"];
+//     }
+//     if (req.body["settings.email.templates.newMessage"]!=undefined) {
+//       update["settings.email.templates.newMessage"] = req.body["settings.email.templates.newMessage"];
+//     }
+//     if (req.body["settings.email.templates.newMessageFollower"]!=undefined) {
+//       update["settings.email.templates.newMessageFollower"] = req.body["settings.email.templates.newMessageFollower"];
+//     }
+//     if (req.body["settings.email.templates.ticket"]!=undefined) {
+//       update["settings.email.templates.ticket"] = req.body["settings.email.templates.ticket"];
+//     }
+//     if (req.body["settings.email.templates.sendTranscript"]!=undefined) {
+//       update["settings.email.templates.sendTranscript"] = req.body["settings.email.templates.sendTranscript"];
+//     }
+//     if (req.body["settings.email.templates.emailDirect"]!=undefined) {
+//       update["settings.email.templates.emailDirect"] = req.body["settings.email.templates.emailDirect"];
+//     }
+//     if (req.body["settings.email.from"]!=undefined) {
+//       update["settings.email.from"] = req.body["settings.email.from"];
+//     }
+//     if (req.body["settings.email.config.host"]!=undefined) {
+//       update["settings.email.config.host"] = req.body["settings.email.config.host"];
+//     }
+//     if (req.body["settings.email.config.port"]!=undefined) {
+//       update["settings.email.config.port"] = req.body["settings.email.config.port"];
+//     }
+//     if (req.body["settings.email.config.secure"]!=undefined) {
+//       update["settings.email.config.secure"] = req.body["settings.email.config.secure"];
+//     }
+//     if (req.body["settings.email.config.user"]!=undefined) {
+//       update["settings.email.config.user"] = req.body["settings.email.config.user"];
+//     }
+//     if (req.body["settings.email.config.pass"]!=undefined) {
+//       update["settings.email.config.pass"] = req.body["settings.email.config.pass"];
+//     }
+//     if (req.body["settings.chat_limit_on"]!=undefined) {
+//       update["settings.chat_limit_on"] = req.body["settings.chat_limit_on"];
+//     }
+//     if (req.body["settings.max_agent_assigned_chat"]!=undefined) {
+//       update["settings.max_agent_assigned_chat"] = req.body["settings.max_agent_assigned_chat"];
+//     }
+//     if (req.body["settings.reassignment_on"]!=undefined) {
+//       update["settings.reassignment_on"] = req.body["settings.reassignment_on"];
+//     }
+//     if (req.body["settings.reassignment_delay"]!=undefined) {
+//       update["settings.reassignment_delay"] = req.body["settings.reassignment_delay"];
+//     }
+//     if (req.body["settings.automatic_unavailable_status_on"]!=undefined) {
+//       update["settings.automatic_unavailable_status_on"] = req.body["settings.automatic_unavailable_status_on"];
+//     }
+//     if (req.body["settings.automatic_idle_chats"]!=undefined) {
+//       update["settings.automatic_idle_chats"] = req.body["settings.automatic_idle_chats"];
+//     }
+
+//     if (req.body.widget!=undefined) {
+//       update.widget = req.body.widget;
+//     }
+//     if (req.body.versions!=undefined) {
+//       update.versions = req.body.versions;
+//     }
+//     if (req.body.channels!=undefined) {
+//       update.channels = req.body.channels; 
+//     }
+//     if (req.body.ipFilterEnabled!=undefined) {
+//       update.ipFilterEnabled = req.body.ipFilterEnabled;
+//     }
+//     if (req.body.ipFilter!=undefined) {
+//       update.ipFilter = req.body.ipFilter;
+//     }
+//     if (req.body.ipFilterDenyEnabled!=undefined) {
+//       update.ipFilterDenyEnabled = req.body.ipFilterDenyEnabled;
+//     }
+//     if (req.body.ipFilterDeny!=undefined) {
+//       update.ipFilterDeny = req.body.ipFilterDeny;
+//     }
+//     if (req.body.bannedUsers!=undefined) {
+//       update.bannedUsers = req.body.bannedUsers;
+//     }
+//     if (req.body.profile!=undefined) {
+//       update.profile = req.body.profile;
+//     }
+
+//     winston.debug('UPDATE PROJECT REQ BODY ', update);
+
+//     Project.findByIdAndUpdate(req.params.projectid, update, { new: true, upsert: true }, function (err, updatedProject) {
+//       if (err) {
+//         winston.error('Error putting project ', err);
+//         return res.status(500).send({ success: false, msg: 'Error updating object.' });
+//       }
+//       projectEvent.emit('project.update', updatedProject );
+//       res.json(updatedProject);
+//     });
+  
+//   } catch (err) {
+//     winston.warn("Profile modification: permission denied.");
+//     res.status(403).send({ success: false, error: "You don't have the permission required to modify the project profile"});
+//   }
+
+// })
+
 router.put('/:projectid', [passport.authenticate(['basic', 'jwt'], { session: false }), validtoken, roleChecker.hasRole('admin')], function (req, res) {
+  
   winston.debug('UPDATE PROJECT REQ BODY ', req.body);
 
   var update = {};
+
+  console.log("ADMIN EMAIL: ", process.env.ADMIN_EMAIL)
+  if (req.body.profile) {
+
+    if (req.user &&
+        req.user.attributes &&
+        req.user.attributes.isSuperadmin === true) {
+          
+          winston.debug("Superadmin can modify the project profile")
+          update.profile = req.body.profile;
+
+          delete req.user.attributes.isSuperadmin;
+        }
+
+        else {
+          winston.verbose("Project profile can't be modified by the current user " + req.user._id);
+          return res.status(403).send({ success: false,  error: "You don't have the permission required to modify the project profile"});
+        }
+
+    // check if super admin
+    // let token = req.headers.authorization
+    // token = token.split(" ")[1];
+
+    // let decoded = jwt.verify(token, configSecret);
+    // winston.debug("user decoded: ", decoded);
+    // console.log("user decoded: ", decoded);
+
+    // if (!process.env.ADMIN_EMAIL) {
+    //   winston.warn("Missing admin email parameter in environment");
+    //   return res.status(401).send({ success: false, error: "Missing admin email parameter"});
+    // }
+
+    // if (!decoded) {
+    //   winston.warn("Profile modification: permission denied.");
+    //   return res.status(403).send({ success: false, error: "You don't have the permission required to modify the project profile. Can't decode user."});
+    // }
+
+    // if (decoded.email !== process.env.ADMIN_EMAIL) {
+    //   winston.warn("Profile modification: permission denied.");
+    //   return res.status(403).send({ success: false, error: "You don't have the permission required to modify the project profile"});
+    // }
+
+    // console.log("You can modify the project profile");
+
+    // winston.info("Illegal field profile detected. Deny project profile update.");
+    // return res.status(403).send({ success: false,  error: "You cannot edit the project profile."});
+  }
   
 //like patch
   if (req.body.name!=undefined) {
@@ -232,10 +457,6 @@ router.put('/:projectid', [passport.authenticate(['basic', 'jwt'], { session: fa
   if (req.body.bannedUsers!=undefined) {
     update.bannedUsers = req.body.bannedUsers;
   }
-
-  if (req.body.profile != undefined) {
-    update.profile = req.body.profile;
-  }
   
   // if (req.body.defaultLanguage!=undefined) {
   //   update.defaultLanguage = req.body.defaultLanguage; 
@@ -243,7 +464,6 @@ router.put('/:projectid', [passport.authenticate(['basic', 'jwt'], { session: fa
 
   
   winston.debug('UPDATE PROJECT REQ BODY ', update);
-
   // console.log("update",JSON.stringify(update));
 
   Project.findByIdAndUpdate(req.params.projectid, update, { new: true, upsert: true }, function (err, updatedProject) {